Fix the type used for the error code for KeychainAccessException.
`size_t` is an unsigned type, so using it was resulting in a large
positive error code being printed for keychain errors, rather than the
real small negative error code.

+++ src/impl/apple/keychain_helper.hpp
@@ -29,7 +29,7 @@ std::vector<char> metadata_realm_encryption_key();
 
 class KeychainAccessException : public std::runtime_error {
 public:
-    KeychainAccessException(size_t error_code);
+    KeychainAccessException(int32_t error_code);
 };
 
 }
